Factor out collections equality comparison
